# swagger/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.indexVideoRequest:
    properties:
      description:
        type: string
      link:
        type: string
    required:
      - link
    type: object
  api.indexVideoResponse:
    properties:
      description:
        type: string
      link:
        type: string
    required:
      - link
    type: object
  api.videoEntry:
    properties:
      description:
        type: string
      link:
        type: string
    type: object
host: api-zvezdolet.ddns.net
info:
  contact: { }
  description: API к сервису индексации и поиска видео
  title: Zvezdolet Search API
  version: "1.0"
paths:
  /index:
    post:
      consumes:
        - application/json
      description: Индексировать видео в базе сервиса
      parameters:
        - description: ссылка и описание видео
          in: body
          name: video
          required: true
          schema:
            $ref: '#/definitions/api.indexVideoRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.indexVideoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Индексировать видео
  /search:
    get:
      description: Поиск видео по заданному тексту
      parameters:
        - description: поисковый запрос
          in: query
          name: text
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.videoEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Поиск видео
swagger: "2.0"